[Fix] #110 - Objective idx int로 타입 변경 및 히스토리 조회시 idx 타입 변경

+++ src/main/java/org/moonshot/server/domain/objective/dto/response/HistoryObjectiveListDto.java
@@ -10,7 +10,7 @@ public record HistoryObjectiveListDto(
         String objCategory,
         int progress,
         String objPeriod,
-        Short objIdx,
+        int objIdx,
         List<HistoryKeyResultDto> krList
 ) {
     public static HistoryObjectiveListDto of(Objective objective) {
